+++ radix-engine/src/engine/mod.rs
@@ -1,14 +1,17 @@
+mod call_frame;
 mod component_objects;
 mod id_allocator;
 mod id_validator;
-mod process;
-pub mod receipt;
+mod receipt;
 mod substate_receipt;
+mod system_api;
 mod track;
 
+pub use call_frame::{CallFrame, SNodeState};
 pub use component_objects::*;
 pub use id_allocator::*;
 pub use id_validator::*;
-pub use process::{Process, SNodeState, SystemApi};
+pub use receipt::*;
 pub use substate_receipt::{CommitReceipt, SubstateOperation, SubstateOperationsReceipt};
+pub use system_api::SystemApi;
 pub use track::{Address, BorrowedSNodes, SubstateParentId, SubstateUpdate, SubstateValue, Track};
